consolidate fill tests and remove redundant beforeEach blocks
- Merge fill-types.test.ts into fill-between.test.ts (renamed from fill-utils.test.ts)
- Remove beforeEach blocks that only clear document.body.innerHTML
  (vitest/happy-dom handles this automatically between tests)
